Moved signalParsers file
It was in wrong folder
